完成自定义负载均衡

+++ FlinkBase/src/main/java/com/wjf/trajectory/FlinkBase/operator/partition/CustomPartitioner.java
@@ -7,6 +7,6 @@ public class CustomPartitioner<T> implements Partitioner<T> {
     @Override
     public int partition(T key, int numPartitions) {
         // Simple partitioning logic for 2 keys and 2 partitions
-        return (Integer) key%numPartitions;
+        return (int) ((Long)key%numPartitions);
     }
 }
